Fix large extent handling

+++ file_scan.h
@@ -47,7 +47,7 @@ struct block_csum {
 struct extent_csum {
 	uint64_t	loff;
 	uint64_t	poff;
-	uint32_t	len;
+	uint64_t	len;
 	unsigned int	flags;
 	unsigned char	digest[DIGEST_LEN];
 };
